[DEP-1832] Implement streaming to better manage memory for NodeJS applications (#10)
Implements and converges this package on streaming, ensuring there is one interface for both the Browser and NodeJS to ensure there is no duplication of logic. This change has been implemented to ensure that NodeJS applications that create large DXFs can do so without running out of memory and it is a deliberate divergence from the forked package.

This change also improves the unit tests by utilising the examples and ensuring they can be regenerated easily. The bundled application within the examples file is now defunct but has been left there.

Includes regenerating the examples from commit #15ff7de4 which was the production commit before this work started


* [DEP-1832] Rework the change to remove duplication (#11)

After discussion we have decided to deviate completely from the fork and
settle on streaming by default with an implementation that still
supports the browser.

This new approach removes all duplication as seen previously.

* only export the types that we export and rename the streamable drawing

* More type fixes